fix: move pool ot core and modify getModuleInfo hook
