+++ packages/core/src/dep/moduleInfoWorker.ts
@@ -1,8 +1,6 @@
 import threads from "worker_threads";
-import { useGetModuleInfo } from "@dep-spy/utils";
+import { getModuleInfo } from "@dep-spy/utils";
 const parentPort = threads.parentPort;
-
-const getModuleInfo = useGetModuleInfo();
 
 parentPort.on("message", async (task: [string, string]) => {
   const moduleInfo = await getModuleInfo(...task);
