+++ packages/core/src/type.ts
@@ -1,3 +1,5 @@
+import { Worker } from "./pool";
+
 export interface Node {
   name: string;
   version: string;
@@ -31,3 +33,12 @@ export interface Config {
     codependency?: string;
   };
 }
+export type Resolve<POOL_TASK extends unknown[], RESULT_TYPE> = (result: {
+  data: RESULT_TYPE;
+  worker: Worker<POOL_TASK, RESULT_TYPE>;
+}) => void;
+
+export type Task<POOL_TASK extends unknown[], RESULT_TYPE> = {
+  task: POOL_TASK;
+  resolve: Resolve<POOL_TASK, RESULT_TYPE>;
+};
