+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
-export { default as useGetModuleInfo } from "./getModuleInfo";
-export { default as Pool, Worker, OnlineWorker, OffLineWorker } from "./pool";
+export { default as getModuleInfo } from "./getModuleInfo";
 export { default as getFileInfo } from "./getFileInfo";
 export type { MODULE_INFO_TYPE } from "./type";
