+++ packages/utils/src/type.ts
@@ -1,5 +1,3 @@
-import { Worker } from "./pool";
-
 export enum INFO_TYPES {
   GITHUB,
   NPM,
@@ -48,13 +46,3 @@ export interface TS_CONFIG {
 }
 
 export type ALIAS_CONFIG = Record<string, string | string[]>;
-
-export type Resolve<POOL_TASK extends unknown[], RESULT_TYPE> = (result: {
-  data: RESULT_TYPE;
-  worker: Worker<POOL_TASK, RESULT_TYPE>;
-}) => void;
-
-export type Task<POOL_TASK extends unknown[], RESULT_TYPE> = {
-  task: POOL_TASK;
-  resolve: Resolve<POOL_TASK, RESULT_TYPE>;
-};
